Extract out st.balloons, latex, text
